windows 64-bit build fix

+++ src/wilton_channel.cpp
@@ -101,7 +101,7 @@ char* wilton_Channel_receive(wilton_Channel* channel, int timeout_millis,
         wilton::support::log_debug(LOGGER, "Receive complete, result: [" + sl::support::to_string_bool(buf.has_value()) + "]");
         if (buf.has_value()) {
             *msg_out = buf.value().data();
-            *msg_len_out = buf.value().size();
+            *msg_len_out = static_cast<int> (buf.value().size());
             *success_out = true;
         } else {
             *success_out = false;
@@ -145,7 +145,7 @@ char* wilton_Channel_poll(wilton_Channel* channel, char** msg_out, int* msg_len_
         wilton::support::log_debug(LOGGER, "Poll complete, result: [" + sl::support::to_string_bool(buf.has_value()) + "]");
         if (buf.has_value()) {
             *msg_out = buf.value().data();
-            *msg_len_out = buf.value().size();
+            *msg_len_out = static_cast<int> (buf.value().size());
             *success_out = true;
         } else {
             *success_out = false;
@@ -169,7 +169,7 @@ char* wilton_Channel_peek(wilton_Channel* channel, char** msg_out, int* msg_len_
         wilton::support::log_debug(LOGGER, "Peek complete, result: [" + sl::support::to_string_bool(buf.has_value()) + "]");
         if (buf.has_value()) {
             *msg_out = buf.value().data();
-            *msg_len_out = buf.value().size();
+            *msg_len_out = static_cast<int> (buf.value().size());
             *success_out = true;
         } else {
             *success_out = false;
